common/requirements.txt
# FastAPI 核心
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# 数据库
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
motor>=3.3.0

# Redis
redis[asyncio]>=5.0.0

# 对象存储
minio>=7.2.0

# Elasticsearch
elasticsearch[async]>=8.11.0,<9.0.0

# HTTP 客户端
httpx>=0.26.0
aiohttp>=3.9.0

# 日志
python-json-logger>=2.0.0

# 文档解析
PyMuPDF>=1.23.0           # fitz：PDF图片提取、页面渲染
pdfplumber>=0.11.0         # PDF表格结构化提取
pdfminer.six>=20221105     # PDF段落级文本提取（含布局感知）
python-docx>=1.1.0         # DOCX 解析（章节树、表格、图片）
openpyxl>=3.1.0            # Excel .xlsx 读写（ragflow 模式精确行数检测）
pandas>=2.0.0              # DataFrame 处理（表格/CSV）
python-pptx>=0.6.23        # PPTX 解析
beautifulsoup4>=4.12.0     # HTML 解析
html5lib>=1.1              # BeautifulSoup html5lib 后端（ragflow HTML 解析器使用）
lxml>=4.9.0                # BeautifulSoup XML/lxml 后端
chardet>=5.2.0             # 文件编码自动检测（ragflow html_parser 使用）
markdown>=3.5.0            # Markdown 转 HTML（MarkdownParser 表格提取）
xxhash>=3.4.0              # Chunk ID 生成（与 ragflow 一致）

# 向量嵌入
sentence-transformers>=2.2.0
torch>=2.0.0

# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# 工具
python-dotenv>=1.0.0
pyyaml>=6.0

# Nacos 服务注册与配置中心
nacos-sdk-python>=2.0.0

# PageIndex 依赖
litellm>=1.83.0
